# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Melodys API",
        "version": "0.5.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/user/password_recovery": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "\u0412\u043e\u0441\u0441\u0442\u0430\u043d\u043e\u0432\u043b\u0435\u043d\u0438\u0435 \u043f\u0430\u0440\u043e\u043b\u044f",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "\u0410\u0434\u0440\u0435\u0441 \u044d\u043b\u0435\u043a\u0442\u0440\u043e\u043d\u043d\u043e\u0439 \u043f\u043e\u0447\u0442\u044b",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/user/password_reset": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "\u0421\u0431\u0440\u043e\u0441 \u043f\u0430\u0440\u043e\u043b\u044f",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "\u0422\u043e\u043a\u0435\u043d",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
...
}